+++ CHANGELOG.md
@@ -10,7 +10,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 - Add loader to profile directory page.
 - Suggest re-verification if identity proof is outdated.
-- Add "Network" timeline.
+- Add "Federated" timeline.
 - Display "Admin" badge if logged in user is admin.
 - Support sign-up notifications.
 - Add "Mute author" to post dropdown menu at local and network timelines.